+++ docs/_build/html/installation.php
@@ -5,7 +5,7 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     
-    <title>Installation &mdash; PyiUpdater 0.15.0-beta-dirty documentation</title>
+    <title>Installation &mdash; PyiUpdater 0.15.1-beta2-1-g7289340-dirty documentation</title>
     
     <link rel="stylesheet" href="_static/basic.css" type="text/css" />
     <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
@@ -15,7 +15,7 @@
     <script type="text/javascript">
       var DOCUMENTATION_OPTIONS = {
         URL_ROOT:    './',
-        VERSION:     '0.15.0-beta-dirty',
+        VERSION:     '0.15.1-beta2-1-g7289340-dirty',
         COLLAPSE_INDEX: false,
         FILE_SUFFIX: '.php',
         HAS_SOURCE:  true
@@ -28,7 +28,7 @@
     <script type="text/javascript" src="_static/js/jquery-fix.js"></script>
     <script type="text/javascript" src="_static/bootstrap-3.2.0/js/bootstrap.min.js"></script>
     <script type="text/javascript" src="_static/bootstrap-sphinx.js"></script>
-    <link rel="top" title="PyiUpdater 0.15.0-beta-dirty documentation" href="index.php" />
+    <link rel="top" title="PyiUpdater 0.15.1-beta2-1-g7289340-dirty documentation" href="index.php" />
     <link rel="next" title="Configuration" href="configuration.php" />
     <link rel="prev" title="Downloads" href="downloads.php" />
 <meta charset='utf-8'>
@@ -50,7 +50,7 @@
         </button>
         <a class="navbar-brand" href="index.php">
           PyiUpdater</a>
-        <span class="navbar-text navbar-version pull-left"><b>0.15.0-beta-dirty</b></span>
+        <span class="navbar-text navbar-version pull-left"><b>0.15.1-beta2-1-g7289340-dirty</b></span>
       </div>
 
         <div class="collapse navbar-collapse nav-collapse">
@@ -144,17 +144,17 @@
 <p>PyiUpdater depends on a few external libraries: <a class="reference external" href="https://pypi.python.org/pypi/appdirs/">appdirs</a>, <a class="reference external" href="https://pypi.python.org/pypi/blinker">blinker</a>, <a class="reference external" href="http://aws.amazon.com/sdkforpython/">boto</a>,  <a class="reference external" href="https://github.com/ilanschnell/bsdiff4">bsdiff4</a>, <a class="reference external" href="https://pypi.python.org/pypi/certifi">certifi</a>, <a class="reference external" href="https://pypi.python.org/pypi/ed25519">ed25519</a>, <a class="reference external" href="https://pypi.python.org/pypi/JMS-Utils">jms_utils</a> , <a class="reference external" href="https://github.com/pyinstaller/pyinstaller">pyinstaller</a>, <a class="reference external" href="https://pypi.python.org/pypi/simple-pbkdf2">simple_pbkdf2</a>, <a class="reference external" href="https://pypi.python.org/pypi/six">six</a>, <a class="reference external" href="https://pypi.python.org/pypi/stevedore">stevedore</a> &amp; <a class="reference external" href="https://pypi.python.org/pypi/urllib3">urllib3</a>. Bsdiff4 is only required to make patches, not to apply them.  These libraries are not documented here.</p>
 <p>So how do you get all that on your computer quickly?</p>
 <p>Install from pip:</p>
-<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater --process-dependency-links
+<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater
 </pre></div>
 </div>
 <p>S3 &amp; SCP upload plugins are available with:</p>
-<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater[s3] --process-dependency-links
+<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater[s3]
 
-$ pip install PyiUpdater[scp] --process-dependency-links
+$ pip install PyiUpdater[scp]
 </pre></div>
 </div>
 <p>If you want the bleeding edge download a pre-release version. WARNING! -&gt; pre released version may not work as expected:</p>
-<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater --pre --process-dependency-links
+<div class="highlight-python"><div class="highlight"><pre>$ pip install PyiUpdater --pre
 </pre></div>
 </div>
 </div>
@@ -172,7 +172,7 @@ $ pip install PyiUpdater[scp] --process-dependency-links
     </p>
     <p>
         &copy; Copyright 2014, Digital Sapphire.<br/>
-      Last updated on Jan 04, 2015.<br/>
+      Last updated on Jan 10, 2015.<br/>
     </p>
   </div>
 </footer>