+++ pyi_updater/hooks/hook-jms_utils.py
@@ -1,0 +1,2 @@
+hiddenimports = ['jms_utils.logger',
+                 'jms_utils.paths',]